C03 feature object3d

+++ Core/DirectX12/TextureManager.cpp
@@ -56,7 +56,7 @@ void TextureManager::LoadTexture(const std::string& _filePath)
     srvDesc.Format = textureData.metadata.format;
     srvDesc.ViewDimension = D3D12_SRV_DIMENSION_TEXTURE2D;
     srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
-    srvDesc.Texture2D.MipLevels = textureData.metadata.mipLevels;
+    srvDesc.Texture2D.MipLevels = static_cast<UINT>(textureData.metadata.mipLevels);
 
     device->CreateShaderResourceView(textureData.resource.Get(), &srvDesc, textureData.srvHandleCPU);
 }
